Model moved to Entity, DbContext formed.
