+++ WebbySoftware/Entity/Interfaces/BaseAtrributes.cs
