+++ WebbySoftware/Entity/Interfaces/ProjectModule.cs
@@ -2,6 +2,7 @@
     public interface ProjectModule {
 
         //Attributes common to all Project Modules
+        public int ID { get; set; }
         public string ProjectName { get; set; }
         public string ProjectDescription { get; set; }
         public string ProjectVersion { get; set; }
